Fix compilation issues on MAC.
	* scribo/core/line_info.hh: Explicitly call parent constructor.
	* scribo/primitive/extract/non_text.hh: Fix ambiguities on
	std::pow overloads.

	* src/debug/show_group_bboxes.cc: Fix namespace ambiguity.

+++ scribo/scribo/primitive/extract/non_text.hh
@@ -116,8 +116,8 @@ namespace scribo
 	  }
 
 
-	  const unsigned q_div = std::pow(2, 8 - nbits);
-	  const unsigned q = unsigned(std::pow(2, nbits));
+	  const unsigned q_div = std::pow(2.f, (int)(8 - nbits));
+	  const unsigned q = unsigned(std::pow(2.f, (int)nbits));
 	  const unsigned nelements = input.nelements();
